rename sample_generator.py to samples_generator.py
