+++ kenchi/datasets/__init__.py
@@ -1,2 +1,2 @@
 from .base import * # noqa
-from .sample_generator import * # noqa
+from .samples_generator import * # noqa
